Herp derp I left the debug in there and forgot to reset errors.

+++ src/init_stack.c
@@ -6,7 +6,7 @@
 /*   By: jcaron <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/28 15:48:49 by jcaron            #+#    #+#             */
-/*   Updated: 2023/01/06 17:28:57 by jcaron           ###   ########.fr       */
+/*   Updated: 2023/01/15 19:22:05 by jcaron           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	init_stack(t_stack *a, t_stack *b, int argc, char **argv)
 
 	tab = arg_to_tab(argc, argv);
 	a->tab = tab;
-	a->top = argc - 1 ;
+	a->top = (size_t)argc - 1 ;
 	b->tab = tab + argc;
-	b->top = argc - 1;
+	b->top = (size_t)argc - 1;
 }
